Detect YAQL version based on the distribution package
YAQL version is now detected based on the package distribution.
If YAQL is not installed via setuptools (or is installed but is run from
different source) the version is undefined.

Package version is temporary hardcoded in setup.cfg as 1.0.0b1, however
it has to be removed before the release: the versions have to be defined
based on the git tags. Right now latest remote tag in master branch is 0.2,
so we have to manually add version to setup.cfg.
As soon as 1.0 tag is added to remote, version field has to be removed.

Change-Id: I5b9a6d724cb5cfc7f33c5d79f5862b4edecc5d27