Make py34 job pass successfully
As py34 job is now voting, no patch to 0.2 branch may land as py34 tests
do not pass in it.

This patch make these tests passable.

NOTE: THIS IS NOT AN INTRODUCTION OF Py34 SUPPORT! YAQL 0.2 STILL DOES
NOT RUN ON Py3x!

This patch will just allow the empty test to properly pass the gate.
Proper py34 support is available on 1.x branch only!

Change-Id: I043b2c0407b25eddc6d527a65d49258a12320219