*args/**kwargs were added to SmartType methods
It is likely that next yaql version may have additional parameters
to check/convert methods. But it is impossible to add parameter
without breaking old inheritors that do not expect those parameters.
The only solution is to add *args/**kwargs to each method from the
start so that old functions will just ignore new parameters

Change-Id: I2b95a5111366ab488a7a9e3223481cc5579395ef