Incorrect behavior of join() on iterators was fixed
Due to the incorrect condition it code that limits
iterators even collections with known limit (lists etc.)
were converted to limiting iterators when there was
no upper length limit. As a result when join() method
was called without iterator limitation being configured
(for example through yaql.eval) it received its collection
arguments as iterators. Because iterators cannot be reset
and re-read again iterator for the inner iterator became
exhausted after the first item of the outer collection.

join() method was fixed to work correctly to memorize
inner iterator so that it could be re-read

Change-Id: I287d0d86b5461490ff32731c11a726174939753d