Ability to yaqlize any Python class was added
Now any Python class or object can be yaqlified
with single function call or decorator.

It is possible to call methods, access attributes/properties
and index yaqlified objects. It is possible to disable any
of these 3 operations and specify whitelist/blacklist
of methods/attributes/keys that are exposed to the
yaql.

Change-Id: I8973492aa6a86d119b05fd33dd8b00ce283fee98