Use __slots__ for yaqltypes
Collectively, these types are fairly common in the context (about 400 in
the default context). Defining __slots__ for them so that they don't
need a __dict__ object to store attributes results in a fairly
inconsequential saving of memory with the default context, but likely
could result in significant savings depending on the amount of data yaql
is operating on.

Change-Id: Iaae6b48672a3293f3920125347835a100883ea76