Use modern PBR and hacking packages
The 2.0.0 is breaking in that it removes the use of warnerrors in
build_sphinux.

YAQL isn't using that feature, so it shoudln't break.

The cap on pbr is preventing other OpenStack projects that would like to
use pbr 2.0.0 (and sphinx 1.5.1) from doing so as it breaks
co-installability with yaql

Also hacking <0.11 had a similar issue so use a newer version of that also

Change-Id: I23e1d7aa715003250131bf5c88a438f278e4d4da
Related-Bug: 1668848