Import docs building job
This repo had a docs building job setup in project-config repo and was
at one time publishing to readthedocs. The publishing needs a manual
change and the repo was never adapted.

Import just the docs building jobs using the build-openstack-docs-pti
template.
If the repo wants to use publish later to readthedocs, the template
can be changed to docs-on-readthedocs as explained in
http://lists.openstack.org/pipermail/openstack-dev/2018-August/132836.html

Change-Id: Idd3725fdc41728d105c1f4d526668673f28b915c